# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API documentation for Recipes Web Shop.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/recipes/{recipe}/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "List ingredients for a recipe",
                "operationId": "c40cd5b1c2c29ccd7a47b80d97a55631",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No ingredients found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Add ingredient to a recipe (admin only)",
                "operationId": "23a8541f33a4ae37e56588cadfa2243e",
                "parameters": [
                    {
                        "name": "recipe",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingredient added"
                    },
                    "403": {
                        "description": "Only admins can add ingredients"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ingredients/{ingredient}": {
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Update an ingredient (admin only)",
                "operationId": "be01c11e6c94a5b354f1c07e4e511655",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Ingredient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ingredient updated"
                    },
                    "403": {
                        "description": "Only admins can update ingredients"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "summary": "Delete an ingredient (admin only)",
                "operationId": "ee34994a8022f644f24d9f73fdcbab5a",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Ingredient ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingredient deleted"
                    },
                    "403": {
                        "description": "Only admins can delete ingredients"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}